Handle token expiration

+++ CroudSeek.Client/Services/QuestDataService.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Blazored.LocalStorage;
 using CroudSeek.Shared;
+using Microsoft.AspNetCore.Components;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -16,7 +17,7 @@ namespace CroudSeek.Client.Services
         private readonly HttpClient _httpClient;
         private readonly IMapper _mapper;
         private readonly IDataPointDataService _dataPointService;
-        public QuestDataService(HttpClient hhttpClient, IMapper mapper, IDataPointDataService dataPointService, IClient client, ILocalStorageService localStorage) : base(client, localStorage)
+        public QuestDataService(HttpClient hhttpClient, IMapper mapper, IDataPointDataService dataPointService, IClient client, ILocalStorageService localStorage, NavigationManager navigation) : base(client, localStorage, navigation)
         {
             _httpClient = client.HttpClient;
             _mapper = mapper;
@@ -24,60 +25,71 @@ namespace CroudSeek.Client.Services
         }
         public async Task<QuestDto> AddQuest(QuestForCreationDto quest)
         {
-            await AddBearerToken();
-
-            var json = JsonSerializer.Serialize(quest);
-
-            var questJson =
-                new StringContent(json, Encoding.UTF8, "application/json");
-            
-            var response = await _httpClient.PostAsync("api/quests", questJson);
-
-            if (response.IsSuccessStatusCode)
+            if (await AddBearerToken())
             {
-                return await JsonSerializer.DeserializeAsync<QuestDto>(await response.Content.ReadAsStreamAsync());
-            }
+                var json = JsonSerializer.Serialize(quest);
 
+                var questJson =
+                    new StringContent(json, Encoding.UTF8, "application/json");
+
+                var response = await _httpClient.PostAsync("api/quests", questJson);
+
+                if (response.IsSuccessStatusCode)
+                {
+                    return await JsonSerializer.DeserializeAsync<QuestDto>(await response.Content.ReadAsStreamAsync());
+                }
+            }
             return null;
         }
 
         public async Task DeleteQuest(int questId)
         {
-            await AddBearerToken();
-            await _httpClient.DeleteAsync($"api/quests/{questId}");
+            if (await AddBearerToken())
+            {
+                await _httpClient.DeleteAsync($"api/quests/{questId}");
+            }
         }
 
         public async Task<IEnumerable<QuestDto>> GetAllQuests()
         {
-            await AddBearerToken();
-            return await JsonSerializer.DeserializeAsync<IEnumerable<QuestDto>>
-                (await _httpClient.GetStreamAsync($"api/quests"), new JsonSerializerOptions() { PropertyNameCaseInsensitive = true });
+            if (await AddBearerToken())
+            {
+                return await JsonSerializer.DeserializeAsync<IEnumerable<QuestDto>>
+                    (await _httpClient.GetStreamAsync($"api/quests"), new JsonSerializerOptions() { PropertyNameCaseInsensitive = true });
+            }
+            return null;
+
         }
 
         public async Task<QuestWithDataPointsDto> GetQuestDetails(int questId)
         {
-            await AddBearerToken();
-            var quest = await JsonSerializer.DeserializeAsync<QuestDto>
-                (await _httpClient.GetStreamAsync($"api/quests/{questId}"), new JsonSerializerOptions() { PropertyNameCaseInsensitive = true });
+            if (await AddBearerToken())
+            {
+                var quest = await JsonSerializer.DeserializeAsync<QuestDto>
+                    (await _httpClient.GetStreamAsync($"api/quests/{questId}"), new JsonSerializerOptions() { PropertyNameCaseInsensitive = true });
 
-            var dpQuest = _mapper.Map<QuestWithDataPointsDto>(quest);
+                var dpQuest = _mapper.Map<QuestWithDataPointsDto>(quest);
 
-            var dataPoints = await _dataPointService.GetDataPointsForQuest(questId);
-            dpQuest.DataPoints = new List<DataPointDto>(dataPoints);
+                var dataPoints = await _dataPointService.GetDataPointsForQuest(questId);
+                dpQuest.DataPoints = new List<DataPointDto>(dataPoints);
 
-            return dpQuest;
+                return dpQuest;
+            }
+            return null;
         }
 
         public async Task UpdateQuest(QuestForUpdateDto quest,int questId)
         {
-            await AddBearerToken();
-            var json = JsonSerializer.Serialize(quest);
+            if (await AddBearerToken())
+            {
+                var json = JsonSerializer.Serialize(quest);
 
-            var questJson =
-                new StringContent(json, Encoding.UTF8, "application/json");
+                var questJson =
+                    new StringContent(json, Encoding.UTF8, "application/json");
 
 
-            var result = await _httpClient.PutAsync($"api/quests/{questId}", questJson);
+                var result = await _httpClient.PutAsync($"api/quests/{questId}", questJson);
+            }
         }
     }
 }
